prefacatoring; extract message building from our send marathons
