+++ orchestrator/cosmos_gravity/src/lib.rs
@@ -9,3 +9,5 @@ extern crate log;
 pub mod query;
 pub mod send;
 pub mod utils;
+
+mod build;
